+++ src/constants/localization.ts
@@ -1,0 +1,9 @@
+const en = {
+  home: 'Home',
+  homeScreen: 'Home Screen',
+  search: 'Search',
+  searchScreen: 'Search Screen',
+  movieDetails: 'Movie Details'
+};
+
+export { en };