+++ src/constants/screenNames.ts
@@ -1,0 +1,7 @@
+export enum screenNames {
+  Home = 'Home',
+  HomeStack = 'HomeStack',
+  MovieDetails = 'MovieDetails',
+  Search = 'Search',
+  BottomTab = 'BottomTab'
+}