+++ CHANGELOG.md
@@ -5,3 +5,10 @@ This file contains all changelogs for the commits.
 ## Commit #1
 
 Setting up eslint, prettier, tsconfig and pre-commit hooks.
+
+## Commit #2
+
+- Navigation setup with Home, Search and Movie Details screen.
+- Setup jest and detox for e2e testing. Using Jest + Testing library for unit and integration test.
+- Internationalization setup for multi-lingual support.
+- Common style utils to avoid re-writing styles.